+++ backend/routes/user.js
@@ -7,6 +7,9 @@ import crypto  from 'crypto';
 import { checkAuth, isValidPassword } from '../utils/auth.js';
 import JobApplicant from '../models/jobApplicant.model.js';
 import JobOffer from '../models/jobOffer.model.js';
+import Report from '../models/report.model.js';
+import Blacklist from '../models/blackList.model.js';
+import jwt from 'jsonwebtoken';
 
 const router = express.Router();
 
@@ -68,6 +71,26 @@ router.post('/user/delete', checkAuth, async (req, res) => {
       return res.status(400).json({ message: 'Invalid credentials' });
     }
 
+    const token = req.headers.authorization?.split(' ')[1];
+  
+    if (!token) {
+      return res.status(400).json({ message: 'Token is required' });
+    }
+  
+    // Decode the token to get its expiration time
+    const decoded = jwt.decode(token);
+  
+    if (!decoded || !decoded.exp) {
+      return res.status(400).json({ message: 'Invalid token' });
+    }
+  
+    const expiryDate = new Date(decoded.exp * 1000); // Convert exp from seconds to milliseconds
+  
+    // Save token in the blacklist with expiration
+    await Blacklist.create({ token, createdAt: expiryDate });
+    await Blacklist.deleteMany({ createdAt: { $lte: new Date() } });
+
+
     await JobApplicant.deleteMany({ applicant: user._id });
     
     await JobOffer.updateMany(
@@ -75,6 +98,13 @@ router.post('/user/delete', checkAuth, async (req, res) => {
       { $pull: { applicants: user._id } }
     );
 
+    await Report.deleteMany({ 
+      $or: [
+        { reportedBy: user._id },
+        { reportedEntity: user._id }
+      ]
+    });
+    
     await User.deleteOne({ email });
 
     res.status(200).json({ message: 'User deleted successfully' });
